# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ApiResponse:
    properties:
      code:
        description: 0 success, non-zero failure
        type: integer
      data:
        description: Optional payload
      message:
        description: Human-readable message
        type: string
    type: object
  storage.createPoolRequest:
    properties:
      auto_disable_threshold:
        type: number
      cloud_config:
        additionalProperties: true
        type: object
      description:
        type: string
      enabled:
        type: boolean
      location:
        description: Pool location URI (required), e.g. local:///absolute/path; type
          derived from scheme
        type: string
      max_size:
        type: integer
      name:
        type: string
      priority:
        type: integer
    required:
    - location
    - max_size
    - name
    type: object
  storage.reconcileRequest:
    properties:
      dry_run:
        type: boolean
      parallel:
        type: integer
      pool_uuid:
        type: string
    type: object
  storage.updatePoolRequest:
    properties:
      auto_disable_threshold:
        type: number
      cloud_config:
        additionalProperties: true
        type: object
      description:
        type: string
      enabled:
        type: boolean
      location:
        type: string
      max_size:
        type: integer
      name:
        type: string
      priority:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: PrismBox backend HTTP API reference
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PrismBox Backend API
  version: "1.0"
paths:
...
  /storage/pools:
    get:
      description: Lists all storage pools with optional filters by type and status
        (authentication required)
      parameters:
      - description: Filter by storage type
        in: query
        name: storage_type
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List storage pools
      tags:
      - Storage
    post:
      consumes:
      - application/json
      description: Creates a new storage pool (authentication required)
      parameters:
      - description: Storage pool payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/storage.createPoolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create storage pool
      tags:
      - Storage
  /storage/pools/{uuid}:
    get:
      description: Returns details for a specific storage pool (authentication required)
      parameters:
      - description: Storage pool UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Storage pool not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get storage pool
      tags:
      - Storage
    patch:
      consumes:
      - application/json
      description: Updates storage pool configuration (authentication required)
      parameters:
      - description: Storage pool UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Update payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/storage.updatePoolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update storage pool
      tags:
      - Storage
  /storage/pools/{uuid}/disable:
    post:
      description: Disables the specified storage pool (authentication required)
      parameters:
      - description: Storage pool UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Operation failed
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Disable storage pool
      tags:
      - Storage
  /storage/pools/{uuid}/enable:
    post:
      description: Enables the specified storage pool (authentication required)
      parameters:
      - description: Storage pool UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Operation failed
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Enable storage pool
      tags:
      - Storage
  /storage/pools/reconcile:
    post:
      consumes:
      - application/json
      description: Triggers reconciliation to verify database records against actual
        storage (authentication required)
      parameters:
      - description: Reconciliation options
        in: body
        name: input
        schema:
          $ref: '#/definitions/storage.reconcileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Operation failed
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Reconcile storage pools
      tags:
      - Storage
  /storage/pools/refresh:
    post:
      description: Refreshes cached information for all storage pools (authentication
        required)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Operation failed
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Refresh storage pool cache
      tags:
      - Storage
  /storage/pools/usage:
    get:
      description: Returns usage comparing database size to actual storage (authentication
        required)
      parameters:
      - description: Optional storage pool UUID; omit for all pools
        in: query
        name: pool_uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get storage pool usage
      tags:
      - Storage
...
securityDefinitions:
  BearerAuth:
    description: 'Use `Authorization: Bearer {token}` or header `x-prismbox-user-token:
      {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"